[SWWEB-2374]: Pesco DOF chart (#1135)
* PESCO - DOF migration

* [SWWEB-2374]: initial commit

* [SWWEB-2374]: preparing 'api' for custom device chart

* [SWWEB-2374]: missing from merge

* [SWWEB-2374]: pesco_device_values query

* [SWWEB-2374]: dof chart init listener

+++ softwrench.sw4.dashboard/classes/service/statistics/StatisticsResponseEntry.cs
@@ -2,13 +2,13 @@
     public class StatisticsResponseEntry {
 
         private readonly string _fieldValue;
-        private readonly int _fieldCount;
+        private readonly long _fieldCount;
         private readonly string _fieldLabel;
 
         /// <param name="fieldValue"></param>
         /// <param name="fieldCount"></param>
         /// <param name="fieldLabel">Value to display (usually a description or formatted label). Defaults to FieldValue if null.</param>
-        public StatisticsResponseEntry(string fieldValue, int fieldCount, string fieldLabel = null) {
+        public StatisticsResponseEntry(string fieldValue, long fieldCount, string fieldLabel = null) {
             _fieldValue = fieldValue;
             _fieldCount = fieldCount;
             _fieldLabel = fieldLabel ?? fieldValue;
@@ -16,7 +16,7 @@
 
         public string FieldValue { get { return _fieldValue; } }
 
-        public int FieldCount { get { return _fieldCount; } }
+        public long FieldCount { get { return _fieldCount; } }
 
         public string FieldLabel { get { return _fieldLabel; } }
     }
